custom aside component with tabs
